Replace ActionResolver, AssertionResolver with StatementResolver
